Reorganize Postgres command registration
